Add support for DynamicSettings, and use this for Custom dims in Slice
A DynamicSetting is one where the type of a setting is known, but
the name(s) of the setting is dependant on the application we are
using. The classic example is the setting of the Custom dimension
members in a post-11.1.2.2 application, where the names and number
of custom dimensions is not known ahead of time.